vulkan: Improve memory manager
The memory manager now allocates chunks of memory on request and then
subcommits then individually.

It's still dumb and has lots of repeated code. It should search first in
the allocations where there's fresh memory, it has to redo the entire
search to find a free space to commit after an allocation. This may add
up for a performance loss on expensive games, but it should be good
enough for now(TM).